pyproject.toml
[tool.poetry]
name = "arcade_twilio"
version = "0.1.0"
description = "A twilio integration to send SMS and WhatsApps."
authors = ["Sergio Serrano <sdserranog@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
arcade-ai = "~0.1.1"
twilio = "^9.3.6"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"

[tool.poetry.dev-dependencies]
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"